charts/apps/llama-vision/app/requirements.txt
gradio>=4.0.0
openai>=1.0.0
pillow>=10.0.0
